Restructure project
* Move: `main.go` to project root
* Move: most of the logic under `cmd` module
* Change: Makefile to address the project structure changes
