+++ cmd/words/embedables/sentences/dorian-gray.json
