+++ cmd/words/embedables/sentences/frankenstein.json
