+++ cmd/words/embedables/words/dorian-gray.json
