+++ cmd/words/embedables/words/frankenstein.json
